# swagger.yml
openapi: 3.0.0
info:
  title: "Shopa API"
  description: "An Api to keep track of grocery shopping"
  version: "1.0.0"
servers:
  - url: "/api"

# The 'components' section is used for holding schemas and other reusable parts of the API specification
...
paths:
...
  /auth/login:
    post: 
      operationId: "users.login_logic"
      tags:
        - users
      summary: "Authenticate user"
      requestBody:
        description: "User credentials"
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "email"
                - "password"
              properties:
                email:
                  type: "string"
                password:
                  type: "string"
      responses:
        "200":
          description: "Successfully authenticated"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  token:
                    type: "string"
        "401":
          description: "Invalid credentials"




# Grocery Creation
...